+++ sdk/resourcemanager/datafactory/armdatafactory/CHANGELOG.md
@@ -1,5 +1,27 @@
 # Release History
 
+## 1.1.0 (2022-05-27)
+### Features Added
+
+- New const `RestServiceAuthenticationTypeOAuth2ClientCredential`
+- New function `GlobalParameterResource.MarshalJSON() ([]byte, error)`
+- New struct `GlobalParameterListResponse`
+- New struct `GlobalParameterResource`
+- New struct `GlobalParametersClientCreateOrUpdateOptions`
+- New struct `GlobalParametersClientCreateOrUpdateResponse`
+- New struct `GlobalParametersClientDeleteOptions`
+- New struct `GlobalParametersClientDeleteResponse`
+- New struct `GlobalParametersClientGetOptions`
+- New struct `GlobalParametersClientGetResponse`
+- New struct `GlobalParametersClientListByFactoryOptions`
+- New struct `GlobalParametersClientListByFactoryResponse`
+- New field `Resource` in struct `RestServiceLinkedServiceTypeProperties`
+- New field `TokenEndpoint` in struct `RestServiceLinkedServiceTypeProperties`
+- New field `ClientSecret` in struct `RestServiceLinkedServiceTypeProperties`
+- New field `ClientID` in struct `RestServiceLinkedServiceTypeProperties`
+- New field `Scope` in struct `RestServiceLinkedServiceTypeProperties`
+
+
 ## 1.0.0 (2022-05-17)
 ### Breaking Changes